+++ test/OrderPool.t.sol
@@ -10,7 +10,8 @@ import "../src/OrderPoolFactory.sol";
 contract OrderPoolTest is Test {
     address constant WETH = 0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2;
     address constant USDC = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
-    address constant ORACLE_ETH_USD = 0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419;
+    address constant ORACLE_ETH_USD =
+        0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419;
 
     IOrderPoolFactory factory;
     IOrderPool pool;
@@ -20,7 +21,12 @@ contract OrderPoolTest is Test {
         console.log("Creator: ", msg.sender);
 
         factory = new OrderPoolFactory();
-        console.log("Order Pool Factory deployed: ", address(factory), " owner: ", factory.owner());
+        console.log(
+            "Order Pool Factory deployed: ",
+            address(factory),
+            " owner: ",
+            factory.owner()
+        );
 
         factory.createPair(ORACLE_ETH_USD, false, WETH, USDC);
         (pool, reversePool) = factory.getPair(WETH, USDC);
@@ -31,8 +37,11 @@ contract OrderPoolTest is Test {
     }
 
     function testPriceFeed() public {
-        uint amount = 10**ERC20(WETH).decimals();
-        assertApproxEqRel(reversePool.convert(pool.convert(amount)), amount, 10e15);
+        uint256 amount = 10**ERC20(WETH).decimals();
+        assertApproxEqRel(
+            reversePool.convert(pool.convert(amount)),
+            amount,
+            10e15
+        );
     }
-
 }
